added some try catch blocks

+++ 09-Capstone/student/java/tenmo-server/src/main/java/com/techelevator/tenmo/dao/UserDAO.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 public interface UserDAO {
 
-	void updatePending(int optionChoice, int transferId);
+	void updatePending(int optionChoice, long transferId);
 
 	void viewTransferDetails(int transferId);
